ai-service/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
tensorflow-cpu==2.10.0
numpy>=1.19.2
matplotlib>=3.3.4
pandas>=1.2.4
pillow>=8.2.0
scikit-learn>=0.24.2
opencv-python>=4.5.2
python-dotenv>=0.19.0
requests>=2.25.1
tqdm>=4.61.0
jupyter>=1.0.0
protobuf==3.20.3
